Updated PR list script

+++ .chezmoiscripts/darwin/run_onchange_after_configure.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
-[ -d "$HOME/bin" ] &&
+# source "$HOME/scripts/helpers.sh"
+
+if [ -d "$HOME/bin" ]; then
     echo "🗑️ Deleting $HOME/bin/chezmoi" && rm -rf "$HOME/bin"
+fi
